convert prices to Naira and move watermark tools to toolsbox folder
