+++ src/app/ui/pages/toolboxpage/watermarking-tool/components/ImageUploadArea.tsx
