+++ src/app/ui/pages/toolboxpage/watermarking-tool/components/PreviewCanvas.tsx
