+++ src/app/ui/pages/toolboxpage/watermarking-tool/components/ui/Slider.tsx
