+++ src/app/ui/pages/toolboxpage/watermarking-tool/hooks/useDragAndDrop.ts
