+++ src/app/ui/pages/toolboxpage/watermarking-tool/styles/index.css
