+++ src/app/ui/pages/toolboxpage/watermarking-tool/utils/canvas-utils.ts
