+++ src/app/ui/pages/toolboxpage/watermarking-tool/utils/image-processing.ts
